Move desig,preSig from urbit/ -> lib/
